replaced alert with modal and moved some code to base page
