+++ src/app/pages/office/cat-b/office.cat-b.page.html
@@ -151,11 +151,13 @@
 </ion-content>
 <ion-footer>
   <ion-row class="mes-full-width-card box-shadow">
-    <ion-col *ngIf="!(pageState.isRekey$ | async)">
-      <ion-button class="ion-float-left" id="defer-button" (click)="defer()">
+    <ion-col *ngIf="!(pageState.isRekey$ | async)" class="ion-no-padding ion-text-left">
+      <ion-button id="defer-button" class="mes-grey-button office-button" (click)="defer()">
         <h3>Save and continue later</h3>
       </ion-button>
-      <ion-button class="mes-primary-button ion-float-right" (click)="onSubmit()">
+    </ion-col>
+    <ion-col *ngIf="!(pageState.isRekey$ | async)" class="ion-no-padding ion-text-end">
+      <ion-button class="mes-primary-button office-button" (click)="onSubmit()">
         <h3>Mark as complete</h3>
       </ion-button>
     </ion-col>
