+++ src/app/pages/office/components/finish-test-modal/finish-test-modal.scss
@@ -4,7 +4,7 @@
   @extend .modal-alert-wrapper;
 
   .modal-alert {
-    width: 450px;
+    width: 400px;
   }
 
   .modal-alert-message {
@@ -15,7 +15,12 @@
     text-align: center;
   }
 
-  .cancel-button {
+  .finish-test-cancel-button {
     color: var(--gds-black);
+  }
+
+  .finish-test-continue-button {
+    color: var(--gds-black);
+    font-weight: bold;
   }
 }
