+++ src/app/shared/classes/test-flow-base-pages/office/office-base-page.ts
@@ -1,6 +1,6 @@
 import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
-import { Platform } from '@ionic/angular';
+import { NavController, Platform } from '@ionic/angular';
 import { Router } from '@angular/router';
 
 import { StoreModel } from '@shared/models/store.model';
@@ -10,6 +10,9 @@ import { AuthenticationProvider } from '@providers/authentication/authentication
 import { ActivityCodeModel } from '@shared/constants/activity-code/activity-code.constants';
 
 import { PracticeableBasePageComponent } from '@shared/classes/practiceable-base-page';
+import { CompleteTest, SavingWriteUpForLater } from '@pages/office/office.actions';
+import { JOURNAL_PAGE } from '@pages/page-names.constants';
+import { PersistTests } from '@store/tests/tests.actions';
 
 export interface CommonOfficePageState {
   activityCode$: Observable<ActivityCodeModel>;
@@ -24,6 +27,7 @@ export abstract class OfficeBasePageComponent extends PracticeableBasePageCompon
     authenticationProvider: AuthenticationProvider,
     router: Router,
     store$: Store<StoreModel>,
+    public navController: NavController,
   ) {
     super(platform, authenticationProvider, router, store$);
   }
@@ -39,6 +43,27 @@ export abstract class OfficeBasePageComponent extends PracticeableBasePageCompon
         select(getActivityCode),
       ),
     };
+  }
+
+  async completeTest() {
+    if (!this.isPracticeMode) {
+      this.store$.dispatch(CompleteTest());
+    }
+    await this.popToRoot();
+  }
+
+  async popToRoot() {
+    if (this.isPracticeMode) {
+      this.exitPracticeMode();
+      return;
+    }
+    await this.navController.navigateBack(JOURNAL_PAGE);
+  }
+
+  async defer() {
+    await this.popToRoot();
+    this.store$.dispatch(SavingWriteUpForLater());
+    this.store$.dispatch(PersistTests());
   }
 
 }
