+++ src/components/common/candidate-section/candidate-section.html
@@ -1,6 +1,6 @@
 <ion-row class="mes-full-width-card" id="candidate-section">
   <ion-col>
-    <h2 id="candidate-name">{{ candidateName }}</h2>
+    <h2 id="candidate-name">Mr Iva Gotta </h2>
     <h3 id="candidate-driver-number">{{ candidateDriverNumber }}</h3>
   </ion-col>
   <ion-col class="ion-align-self-end">
